More Hours and mins inputs added
The new hours and mins inputs are now added to the other 3 html files.
Also a change was made in the css file with opacity set to 1 to the
disabled inputs so they don't show up in lightgrey on iphones and ipads.
This needs to be checked on an iphone to see if it has worked.